Make audio dequeuing a CircularBuffer's responsibility
